add checks to install.sh ...

+++ install.sh
@@ -2,14 +2,15 @@
 
 if [ -e "~/.music_shell/cache" ]
     md -p ~/.music_shell/cache
+    touch ~/.music_shell/queue
+    touch ~/.music_shell/config.txt
 fi
 
 if [ -e "~/bin" ]; then
     md ~/bin
+    cp -rf ../src/* ~/bin
 fi
 
-cp -rf ../src/* ~/bin
-
-ln ./def_search.sh s
-touch ~/.music_shell/queue
-touch ~/.music_shell/config.txt
+if [ -e "./def_search.sh" ]
+    ln ./def_search.sh s
+fi
